refactor(package): change setup name into 'game'
